requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
requests>=2.31.0
watchdog>=4.0.0